Migrate real_property and baltimore_msa_streets into external cache
Loading the real_property dataset was creating significant performance problems for the package. The data could not be moved to inst/extdata because the uncompressed spatial data files exceeded the 100 MB size limit for GitHub. This change:

- Implements extdata parameter in get_area_data function
- Update map_area_property and get_area_property functions to access cached data with a wkt spatial filter

This change does not create a way for the user to download, process, and store the cached data.